refactor: extract ViewportMgr to slick.core.ts with contracts in models (Phase 4, milestone 17)
Moves the ViewportMgr class (~1,500 lines) out of slick.grid.ts into slick.core.ts
— the one file every script-tag page already loads before the grid — using the
established cross-file pattern: exported from core, registered on the SlickCore
namespace object (alongside DragExtendHandle, the live precedent for a grid-support
class in core; TreeColumns is the historical one), consumed by the grid via
'import { ViewportMgr as ViewportMgr_ }' + 'IIFE_ONLY ? Slick.ViewportMgr :
ViewportMgr_'. The five geometry/state interfaces move to
src/models/viewportMgr.interface.ts per house convention (type-only, excluded from
iife builds), re-exported from models/index.ts; the Slick global declaration in
global.d.ts gains the member.

Zero consumer impact: no new script tag (a standalone slick.viewportmgr.js would
have broken script-tag pages — the Phase 1 rationale, now resolved by core
placement), DOM byte-identical, esm/cjs pick the export up via index.ts.
ViewportMgr is now independently importable (e.g. for future unit tests).

Verified: iife smoke test (Slick.ViewportMgr resolves and instantiates like
DragExtendHandle), gates across script-tag and esm pages, and full suite green
BEFORE this commit: 632 tests, 631 pass / 1 pending.
