+++ src/slick.grid.ts
@@ -90,6 +90,7 @@ import {
   SlickEventData as SlickEventData_,
   SlickRange as SlickRange_,
   Utils as Utils_,
+  ViewportMgr as ViewportMgr_,
   SelectionUtils as SelectionUtils_,
   ValueFilterMode as ValueFilterMode_,
   WidthEvalMode as WidthEvalMode_,
@@ -117,6 +118,7 @@ const Draggable = IIFE_ONLY ? Slick.Draggable : Draggable_;
 const MouseWheel = IIFE_ONLY ? Slick.MouseWheel : MouseWheel_;
 const Resizable = IIFE_ONLY ? Slick.Resizable : Resizable_;
 const DragExtendHandle = IIFE_ONLY ? Slick.DragExtendHandle : DragExtendHandle_;
+const ViewportMgr = IIFE_ONLY ? Slick.ViewportMgr : ViewportMgr_;
 
 /**
  * @license
@@ -144,1517 +146,6 @@ interface RowCaching {
   cellColSpans: Array<number | '*'>;
   cellNodesByColumnIdx: HTMLElement[];
   cellRenderQueue: any[];
-}
-
-/** Snapshot of the grid's freeze configuration, pushed into ViewportMgr by setFrozenOptions(). */
-interface ViewportFreezeState {
-  frozenColumnIdx: number;
-  hasFrozenRows: boolean;
-  actualFrozenRow: number;
-  frozenBottom: boolean;
-  /** the frozenRow option value — number of rows in the frozen row band (0/-1 when none) */
-  frozenRowCount?: number;
-  /** the frozenRightColumn option value — number of columns frozen at the right edge (0 when none) */
-  frozenRightColCount?: number;
-  /** index of the first right-frozen column (columns.length when the band is off) */
-  frozenRightStartIdx?: number;
-  /** the frozenBottomRow option value — rows frozen at the bottom ALONGSIDE top rows (0 when none) */
-  frozenBottomRowCount?: number;
-  /** first row of the bottom-frozen band = dataLength − frozenBottomRow (MAX_SAFE_INTEGER when off) */
-  bottomFrozenSplitRow?: number;
-}
-
-/**
- * Band-count view of the freeze configuration (Phase 4 groundwork for the 3×3 band
- * model): a zero count means the band does not exist. Derived by updateFreezeState
- * from the legacy freeze snapshot; frozenRightCols stays 0 until right-frozen
- * columns land.
- */
-interface FreezeBandCounts {
-  frozenLeftCols: number;
-  frozenRightCols: number;
-  frozenTopRows: number;
-  frozenBottomRows: number;
-}
-
-/** Geometry inputs for ViewportMgr.applyCanvasWidths — computed by the grid, distributed by the manager. */
-interface CanvasWidthsGeometry {
-  widthChanged: boolean;
-  canvasWidth: number;
-  canvasWidthL: number;
-  canvasWidthR: number;
-  canvasWidthRF: number;
-  headersWidthL: number;
-  headersWidthR: number;
-  headersWidthRF: number;
-  viewportW: number;
-  viewportHasVScroll: boolean;
-  scrollbarWidth: number;
-  createFooterRow?: boolean;
-  createPreHeaderPanel?: boolean;
-  preHeaderPanelWidth?: number | string;
-}
-
-/** Geometry inputs for ViewportMgr.applyPaneHeights — computed by the grid, distributed by the manager. */
-interface PaneHeightsGeometry {
-  viewportH: number;
-  frozenRowsHeight: number;
-  /** height of the bottom-frozen row band (simultaneous top+bottom mode; 0 otherwise) */
-  frozenBottomRowsHeight?: number;
-  scrollbarHeight: number;
-  topPanelH: number;
-  headerRowH: number;
-  footerRowH: number;
-  /** lazily computed to avoid an unconditional style recalc; only read on the autoHeight+frozen path */
-  getContainerVBoxDelta: () => number;
-  autoHeight?: boolean;
-  showPreHeaderPanel?: boolean;
-  preHeaderPanelHeight?: number;
-  showTopHeaderPanel?: boolean;
-  topHeaderPanelHeight?: number;
-  showHeaderRow?: boolean;
-  headerRowHeight?: number;
-}
-
-/** Options consumed by ViewportMgr when constructing the pane/viewport/canvas DOM. */
-interface ViewportMgrBuildOptions {
-  createPreHeaderPanel?: boolean;
-  showPreHeaderPanel?: boolean;
-  createFooterRow?: boolean;
-  showFooterRow?: boolean;
-  showColumnHeader?: boolean;
-  showTopPanel?: boolean;
-  showHeaderRow?: boolean;
-  viewportClass?: string;
-  lazyPanes?: boolean;
-  frozenColumn?: number;
-  frozenRow?: number;
-}
-
-/**
- * ViewportMgr — owns the construction of the grid's pane/viewport/canvas DOM.
- *
- * Phase 1 of the frozen rows/columns encapsulation refactor: this class builds the
- * exact same 6-pane / 4-viewport / 4-canvas structure the grid has always built
- * (characterized by cypress/e2e/dom-shape-characterization.cy.ts) and SlickGrid keeps
- * aliases to every element, so all existing logic is unchanged. Later phases move pane
- * selection, geometry distribution and scroll synchronization in here.
- */
-class ViewportMgr {
-  /** the grid container, captured by buildPanes */
-  protected container!: HTMLElement;
-
-  /**
-   * True when the grid opted into lazyPanes AND no rows/columns were frozen at build
-   * time — only the top-left pane set exists until freezing is enabled.
-   */
-  protected lazy = false;
-
-  /**
-   * Array positions of dynamically materialized viewports/canvases (the viewport and
-   * canvas arrays always extend in lockstep, so one slot serves both). Recorded at
-   * materialization time instead of hardcoding, so band order never matters.
-   */
-  protected rfTopSlot = -1;
-  protected rfBottomSlot = -1;
-  protected bfSlotL = -1;
-  protected bfSlotR = -1;
-  protected bfSlotRF = -1;
-
-  // panes
-  paneHeaderL!: HTMLDivElement;
-  paneHeaderR!: HTMLDivElement;
-  paneTopL!: HTMLDivElement;
-  paneTopR!: HTMLDivElement;
-  paneBottomL!: HTMLDivElement;
-  paneBottomR!: HTMLDivElement;
-
-  // pre-header panels (only when createPreHeaderPanel)
-  preHeaderPanelScroller!: HTMLDivElement;
-  preHeaderPanel!: HTMLDivElement;
-  preHeaderPanelSpacer!: HTMLDivElement;
-  preHeaderPanelScrollerR!: HTMLDivElement;
-  preHeaderPanelR!: HTMLDivElement;
-  preHeaderPanelSpacerR!: HTMLDivElement;
-
-  // header scrollers and header column containers
-  headerScrollerL!: HTMLDivElement;
-  headerScrollerR!: HTMLDivElement;
-  headerScroller: HTMLDivElement[] = [];
-  headerL!: HTMLDivElement;
-  headerR!: HTMLDivElement;
-  headers: HTMLDivElement[] = [];
-
-  // header rows
-  headerRowScrollerL!: HTMLDivElement;
-  headerRowScrollerR!: HTMLDivElement;
-  headerRowScroller: HTMLDivElement[] = [];
-  headerRowSpacerL!: HTMLDivElement;
-  headerRowSpacerR!: HTMLDivElement;
-  headerRowL!: HTMLDivElement;
-  headerRowR!: HTMLDivElement;
-  headerRows: HTMLDivElement[] = [];
-
-  // top panels
-  topPanelScrollerL!: HTMLDivElement;
-  topPanelScrollerR!: HTMLDivElement;
-  topPanelScrollers: HTMLDivElement[] = [];
-  topPanelL!: HTMLDivElement;
-  topPanelR!: HTMLDivElement;
-  topPanels: HTMLDivElement[] = [];
-
-  // viewports and canvases
-  viewportTopL!: HTMLDivElement;
-  viewportTopR!: HTMLDivElement;
-  viewportBottomL!: HTMLDivElement;
-  viewportBottomR!: HTMLDivElement;
-  viewport: HTMLDivElement[] = [];
-  canvasTopL!: HTMLDivElement;
-  canvasTopR!: HTMLDivElement;
-  canvasBottomL!: HTMLDivElement;
-  canvasBottomR!: HTMLDivElement;
-  canvas: HTMLDivElement[] = [];
-
-  // right-frozen band (Phase 4 — exists only while frozenRightColumn > 0 has been applied)
-  paneHeaderRF!: HTMLDivElement;
-  paneTopRF!: HTMLDivElement;
-  paneBottomRF!: HTMLDivElement;
-  headerScrollerRF!: HTMLDivElement;
-  headerRF!: HTMLDivElement;
-  headerRowScrollerRF!: HTMLDivElement;
-  headerRowSpacerRF!: HTMLDivElement;
-  headerRowRF!: HTMLDivElement;
-  topPanelScrollerRF!: HTMLDivElement;
-  topPanelRF!: HTMLDivElement;
-  viewportTopRF!: HTMLDivElement;
-  viewportBottomRF!: HTMLDivElement;
-  canvasTopRF!: HTMLDivElement;
-  canvasBottomRF!: HTMLDivElement;
-  footerRowScrollerRF!: HTMLDivElement;
-  footerRowSpacerRF!: HTMLDivElement;
-  footerRowRF!: HTMLDivElement;
-
-  // bottom-frozen row band (Phase 4 — exists only in simultaneous top+bottom mode)
-  paneBottomFrozenL!: HTMLDivElement;
-  paneBottomFrozenR!: HTMLDivElement;
-  paneBottomFrozenRF!: HTMLDivElement;
-  viewportBottomFrozenL!: HTMLDivElement;
-  viewportBottomFrozenR!: HTMLDivElement;
-  viewportBottomFrozenRF!: HTMLDivElement;
-  canvasBottomFrozenL!: HTMLDivElement;
-  canvasBottomFrozenR!: HTMLDivElement;
-  canvasBottomFrozenRF!: HTMLDivElement;
-
-  // footer rows (only when createFooterRow)
-  footerRowScrollerL!: HTMLDivElement;
-  footerRowScrollerR!: HTMLDivElement;
-  footerRowScroller: HTMLDivElement[] = [];
-  footerRowSpacerL!: HTMLDivElement;
-  footerRowSpacerR!: HTMLDivElement;
-  footerRowL!: HTMLDivElement;
-  footerRowR!: HTMLDivElement;
-  footerRow: HTMLDivElement[] = [];
-
-  /**
-   * Builds the pane/viewport/canvas DOM inside the given container.
-   * The construction order and every class/style is identical to the historical
-   * inline construction in SlickGrid.initialize().
-   */
-  buildPanes(container: HTMLElement, o: ViewportMgrBuildOptions) {
-    this.container = container;
-    this.lazy = !!o.lazyPanes && !(((o.frozenColumn ?? -1) > -1) || ((o.frozenRow ?? -1) > -1));
-
-    // Containers used for scrolling frozen columns and rows.
-    // Under lazyPanes with nothing frozen, only the top-left pane set is built;
-    // the creation ORDER of the conditional elements must stay canonical so both
-    // modes produce the same sibling sequence for whatever exists.
-    this.paneHeaderL = Utils.createDomElement('div', { className: 'slick-pane slick-pane-header slick-pane-left', tabIndex: 0 }, container);
-    if (!this.lazy) {
-      this.paneHeaderR = Utils.createDomElement('div', { className: 'slick-pane slick-pane-header slick-pane-right', tabIndex: 0 }, container);
-    }
-    this.paneTopL = Utils.createDomElement('div', { className: 'slick-pane slick-pane-top slick-pane-left', tabIndex: 0 }, container);
-    if (!this.lazy) {
-      this.paneTopR = Utils.createDomElement('div', { className: 'slick-pane slick-pane-top slick-pane-right', tabIndex: 0 }, container);
-      this.paneBottomL = Utils.createDomElement('div', { className: 'slick-pane slick-pane-bottom slick-pane-left', tabIndex: 0 }, container);
-      this.paneBottomR = Utils.createDomElement('div', { className: 'slick-pane slick-pane-bottom slick-pane-right', tabIndex: 0 }, container);
-    }
-
-    if (o.createPreHeaderPanel) {
-      this.preHeaderPanelScroller = Utils.createDomElement('div', { className: 'slick-preheader-panel ui-state-default slick-state-default', style: { overflow: 'hidden', position: 'relative' } }, this.paneHeaderL);
-      this.preHeaderPanelScroller.appendChild(document.createElement('div'));
-      this.preHeaderPanel = Utils.createDomElement('div', null, this.preHeaderPanelScroller);
-      this.preHeaderPanelSpacer = Utils.createDomElement('div', { style: { display: 'block', height: '1px', position: 'absolute', top: '0px', left: '0px' } }, this.preHeaderPanelScroller);
-
-      if (!this.lazy) {
-        this.preHeaderPanelScrollerR = Utils.createDomElement('div', { className: 'slick-preheader-panel ui-state-default slick-state-default', style: { overflow: 'hidden', position: 'relative' } }, this.paneHeaderR);
-        this.preHeaderPanelR = Utils.createDomElement('div', null, this.preHeaderPanelScrollerR);
-        this.preHeaderPanelSpacerR = Utils.createDomElement('div', { style: { display: 'block', height: '1px', position: 'absolute', top: '0px', left: '0px' } }, this.preHeaderPanelScrollerR);
-      }
-
-      if (!o.showPreHeaderPanel) {
-        Utils.hide(this.preHeaderPanelScroller);
-        this.hideIf(this.preHeaderPanelScrollerR);
-      }
-    }
-
-    // Append the header scroller containers
-    this.headerScrollerL = Utils.createDomElement('div', { className: 'slick-header ui-state-default slick-state-default slick-header-left' }, this.paneHeaderL);
-    if (!this.lazy) {
-      this.headerScrollerR = Utils.createDomElement('div', { className: 'slick-header ui-state-default slick-state-default slick-header-right' }, this.paneHeaderR);
-    }
-
-    // Cache the header scroller containers
-    this.headerScroller.push(this.headerScrollerL);
-    if (!this.lazy) {
-      this.headerScroller.push(this.headerScrollerR);
-    }
-
-    // Append the columnn containers to the headers
-    this.headerL = Utils.createDomElement('div', { className: 'slick-header-columns slick-header-columns-left', role: 'row', style: { left: '-1000px' } }, this.headerScrollerL);
-    if (!this.lazy) {
-      this.headerR = Utils.createDomElement('div', { className: 'slick-header-columns slick-header-columns-right', role: 'row', style: { left: '-1000px' } }, this.headerScrollerR);
-    }
-
-    // Cache the header columns
-    this.headers = this.lazy ? [this.headerL] : [this.headerL, this.headerR];
-
-    this.headerRowScrollerL = Utils.createDomElement('div', { className: 'slick-headerrow ui-state-default slick-state-default' }, this.paneTopL);
-    if (!this.lazy) {
-      this.headerRowScrollerR = Utils.createDomElement('div', { className: 'slick-headerrow ui-state-default slick-state-default' }, this.paneTopR);
-    }
-
-    this.headerRowScroller = this.lazy ? [this.headerRowScrollerL] : [this.headerRowScrollerL, this.headerRowScrollerR];
-
-    this.headerRowSpacerL = Utils.createDomElement('div', { style: { display: 'block', height: '1px', position: 'absolute', top: '0px', left: '0px' } }, this.headerRowScrollerL);
-    if (!this.lazy) {
-      this.headerRowSpacerR = Utils.createDomElement('div', { style: { display: 'block', height: '1px', position: 'absolute', top: '0px', left: '0px' } }, this.headerRowScrollerR);
-    }
-
-    this.headerRowL = Utils.createDomElement('div', { className: 'slick-headerrow-columns slick-headerrow-columns-left' }, this.headerRowScrollerL);
-    if (!this.lazy) {
-      this.headerRowR = Utils.createDomElement('div', { className: 'slick-headerrow-columns slick-headerrow-columns-right' }, this.headerRowScrollerR);
-    }
-
-    this.headerRows = this.lazy ? [this.headerRowL] : [this.headerRowL, this.headerRowR];
-
-    // Append the top panel scroller
-    this.topPanelScrollerL = Utils.createDomElement('div', { className: 'slick-top-panel-scroller ui-state-default slick-state-default' }, this.paneTopL);
-    if (!this.lazy) {
-      this.topPanelScrollerR = Utils.createDomElement('div', { className: 'slick-top-panel-scroller ui-state-default slick-state-default' }, this.paneTopR);
-    }
-
-    this.topPanelScrollers = this.lazy ? [this.topPanelScrollerL] : [this.topPanelScrollerL, this.topPanelScrollerR];
-
-    // Append the top panel
-    this.topPanelL = Utils.createDomElement('div', { className: 'slick-top-panel', style: { width: '10000px' } }, this.topPanelScrollerL);
-    if (!this.lazy) {
-      this.topPanelR = Utils.createDomElement('div', { className: 'slick-top-panel', style: { width: '10000px' } }, this.topPanelScrollerR);
-    }
-
-    this.topPanels = this.lazy ? [this.topPanelL] : [this.topPanelL, this.topPanelR];
-
-    if (!o.showColumnHeader) {
-      this.headerScroller.forEach((el) => {
-        Utils.hide(el);
-      });
-    }
-
-    if (!o.showTopPanel) {
-      this.topPanelScrollers.forEach((scroller) => {
-        Utils.hide(scroller);
-      });
-    }
-
-    if (!o.showHeaderRow) {
-      this.headerRowScroller.forEach((scroller) => {
-        Utils.hide(scroller);
-      });
-    }
-
-    // Append the viewport containers
-    this.viewportTopL = Utils.createDomElement('div', { className: 'slick-viewport slick-viewport-top slick-viewport-left', tabIndex: 0 }, this.paneTopL);
-    if (!this.lazy) {
-      this.viewportTopR = Utils.createDomElement('div', { className: 'slick-viewport slick-viewport-top slick-viewport-right', tabIndex: 0 }, this.paneTopR);
-      this.viewportBottomL = Utils.createDomElement('div', { className: 'slick-viewport slick-viewport-bottom slick-viewport-left', tabIndex: 0 }, this.paneBottomL);
-      this.viewportBottomR = Utils.createDomElement('div', { className: 'slick-viewport slick-viewport-bottom slick-viewport-right', tabIndex: 0 }, this.paneBottomR);
-    }
-
-    // Cache the viewports
-    this.viewport = this.lazy
-      ? [this.viewportTopL]
-      : [this.viewportTopL, this.viewportTopR, this.viewportBottomL, this.viewportBottomR];
-    if (o.viewportClass) {
-      this.viewport.forEach((view) => {
-        view.classList.add(...Utils.classNameToList((o.viewportClass)));
-      });
-    }
-
-    // Append the canvas containers
-    this.canvasTopL = Utils.createDomElement('div', { className: 'grid-canvas grid-canvas-top grid-canvas-left', tabIndex: 0 }, this.viewportTopL);
-    if (!this.lazy) {
-      this.canvasTopR = Utils.createDomElement('div', { className: 'grid-canvas grid-canvas-top grid-canvas-right', tabIndex: 0 }, this.viewportTopR);
-      this.canvasBottomL = Utils.createDomElement('div', { className: 'grid-canvas grid-canvas-bottom grid-canvas-left', tabIndex: 0 }, this.viewportBottomL);
-      this.canvasBottomR = Utils.createDomElement('div', { className: 'grid-canvas grid-canvas-bottom grid-canvas-right', tabIndex: 0 }, this.viewportBottomR);
-    }
-
-    // Cache the canvases
-    this.canvas = this.lazy
-      ? [this.canvasTopL]
-      : [this.canvasTopL, this.canvasTopR, this.canvasBottomL, this.canvasBottomR];
-  }
-
-  /**
-   * Builds the footer-row containers (only called when the createFooterRow option is on).
-   * Identical construction to the historical inline code, including the R-before-L
-   * scroller creation order and spacer widths.
-   */
-  buildFooterRows(o: ViewportMgrBuildOptions, canvasWithScrollbarWidth: number) {
-    if (!this.lazy) {
-      this.footerRowScrollerR = Utils.createDomElement('div', { className: 'slick-footerrow ui-state-default slick-state-default' }, this.paneTopR);
-    }
-    this.footerRowScrollerL = Utils.createDomElement('div', { className: 'slick-footerrow ui-state-default slick-state-default' }, this.paneTopL);
-
-    this.footerRowScroller = this.lazy ? [this.footerRowScrollerL] : [this.footerRowScrollerL, this.footerRowScrollerR];
-
-    this.footerRowSpacerL = Utils.createDomElement('div', { style: { display: 'block', height: '1px', position: 'absolute', top: '0px', left: '0px' } }, this.footerRowScrollerL);
-    Utils.width(this.footerRowSpacerL, canvasWithScrollbarWidth);
-    if (!this.lazy) {
-      this.footerRowSpacerR = Utils.createDomElement('div', { style: { display: 'block', height: '1px', position: 'absolute', top: '0px', left: '0px' } }, this.footerRowScrollerR);
-      Utils.width(this.footerRowSpacerR, canvasWithScrollbarWidth);
-    }
-
-    this.footerRowL = Utils.createDomElement('div', { className: 'slick-footerrow-columns slick-footerrow-columns-left' }, this.footerRowScrollerL);
-    if (!this.lazy) {
-      this.footerRowR = Utils.createDomElement('div', { className: 'slick-footerrow-columns slick-footerrow-columns-right' }, this.footerRowScrollerR);
-    }
-
-    this.footerRow = this.lazy ? [this.footerRowL] : [this.footerRowL, this.footerRowR];
-
-    if (!o.showFooterRow) {
-      this.footerRowScroller.forEach((scroller) => {
-        Utils.hide(scroller);
-      });
-    }
-  }
-
-  /**
-   * Builds the right/bottom panes, chrome, viewports and canvases that a lazyPanes
-   * grid skipped at init, inserting each pane at its canonical sibling position and
-   * pushing the new elements into the shared caches IN PLACE (the grid's array
-   * aliases keep working). Idempotent: returns false when the grid is not lazy
-   * (already fully built or built non-lazy).
-   */
-  materializeSecondaryPanes(o: ViewportMgrBuildOptions): boolean {
-    if (!this.lazy) {
-      return false;
-    }
-    this.lazy = false;
-
-    const container = this.container;
-
-    // panes, at their canonical sibling positions
-    this.paneHeaderR = Utils.createDomElement('div', { className: 'slick-pane slick-pane-header slick-pane-right', tabIndex: 0 });
-    container.insertBefore(this.paneHeaderR, this.paneTopL);
-    this.paneTopR = Utils.createDomElement('div', { className: 'slick-pane slick-pane-top slick-pane-right', tabIndex: 0 });
-    container.insertBefore(this.paneTopR, this.paneTopL.nextSibling);
-    this.paneBottomL = Utils.createDomElement('div', { className: 'slick-pane slick-pane-bottom slick-pane-left', tabIndex: 0 });
-    container.insertBefore(this.paneBottomL, this.paneTopR.nextSibling);
-    this.paneBottomR = Utils.createDomElement('div', { className: 'slick-pane slick-pane-bottom slick-pane-right', tabIndex: 0 });
-    container.insertBefore(this.paneBottomR, this.paneBottomL.nextSibling);
-
-    if (o.createPreHeaderPanel) {
-      this.preHeaderPanelScrollerR = Utils.createDomElement('div', { className: 'slick-preheader-panel ui-state-default slick-state-default', style: { overflow: 'hidden', position: 'relative' } }, this.paneHeaderR);
-      this.preHeaderPanelR = Utils.createDomElement('div', null, this.preHeaderPanelScrollerR);
-      this.preHeaderPanelSpacerR = Utils.createDomElement('div', { style: { display: 'block', height: '1px', position: 'absolute', top: '0px', left: '0px' } }, this.preHeaderPanelScrollerR);
-
-      if (!o.showPreHeaderPanel) {
-        Utils.hide(this.preHeaderPanelScrollerR);
-      }
-    }
-
-    // header scroller + header columns
-    this.headerScrollerR = Utils.createDomElement('div', { className: 'slick-header ui-state-default slick-state-default slick-header-right' }, this.paneHeaderR);
-    this.headerScroller.push(this.headerScrollerR);
-    this.headerR = Utils.createDomElement('div', { className: 'slick-header-columns slick-header-columns-right', role: 'row', style: { left: '-1000px' } }, this.headerScrollerR);
-    this.headers.push(this.headerR);
-
-    // header row
-    this.headerRowScrollerR = Utils.createDomElement('div', { className: 'slick-headerrow ui-state-default slick-state-default' }, this.paneTopR);
-    this.headerRowScroller.push(this.headerRowScrollerR);
-    this.headerRowSpacerR = Utils.createDomElement('div', { style: { display: 'block', height: '1px', position: 'absolute', top: '0px', left: '0px' } }, this.headerRowScrollerR);
-    this.headerRowR = Utils.createDomElement('div', { className: 'slick-headerrow-columns slick-headerrow-columns-right' }, this.headerRowScrollerR);
-    this.headerRows.push(this.headerRowR);
-
-    // top panel
-    this.topPanelScrollerR = Utils.createDomElement('div', { className: 'slick-top-panel-scroller ui-state-default slick-state-default' }, this.paneTopR);
-    this.topPanelScrollers.push(this.topPanelScrollerR);
-    this.topPanelR = Utils.createDomElement('div', { className: 'slick-top-panel', style: { width: '10000px' } }, this.topPanelScrollerR);
-    this.topPanels.push(this.topPanelR);
-
-    if (!o.showColumnHeader) {
-      Utils.hide(this.headerScrollerR);
-    }
-    if (!o.showTopPanel) {
-      Utils.hide(this.topPanelScrollerR);
-    }
-    if (!o.showHeaderRow) {
-      Utils.hide(this.headerRowScrollerR);
-    }
-
-    // viewports (pushed in canonical [TopL, TopR, BottomL, BottomR] order)
-    this.viewportTopR = Utils.createDomElement('div', { className: 'slick-viewport slick-viewport-top slick-viewport-right', tabIndex: 0 }, this.paneTopR);
-    this.viewportBottomL = Utils.createDomElement('div', { className: 'slick-viewport slick-viewport-bottom slick-viewport-left', tabIndex: 0 }, this.paneBottomL);
-    this.viewportBottomR = Utils.createDomElement('div', { className: 'slick-viewport slick-viewport-bottom slick-viewport-right', tabIndex: 0 }, this.paneBottomR);
-    this.viewport.push(this.viewportTopR, this.viewportBottomL, this.viewportBottomR);
-    if (o.viewportClass) {
-      const viewportClassList = Utils.classNameToList(o.viewportClass);
-      this.viewportTopR.classList.add(...viewportClassList);
-      this.viewportBottomL.classList.add(...viewportClassList);
-      this.viewportBottomR.classList.add(...viewportClassList);
-    }
-
-    // canvases
-    this.canvasTopR = Utils.createDomElement('div', { className: 'grid-canvas grid-canvas-top grid-canvas-right', tabIndex: 0 }, this.viewportTopR);
-    this.canvasBottomL = Utils.createDomElement('div', { className: 'grid-canvas grid-canvas-bottom grid-canvas-left', tabIndex: 0 }, this.viewportBottomL);
-    this.canvasBottomR = Utils.createDomElement('div', { className: 'grid-canvas grid-canvas-bottom grid-canvas-right', tabIndex: 0 }, this.viewportBottomR);
-    this.canvas.push(this.canvasTopR, this.canvasBottomL, this.canvasBottomR);
-
-    // footer row (right side; the left one was built at init when createFooterRow)
-    if (o.createFooterRow) {
-      this.footerRowScrollerR = Utils.createDomElement('div', { className: 'slick-footerrow ui-state-default slick-state-default' }, this.paneTopR);
-      this.footerRowScroller.push(this.footerRowScrollerR);
-      this.footerRowSpacerR = Utils.createDomElement('div', { style: { display: 'block', height: '1px', position: 'absolute', top: '0px', left: '0px' } }, this.footerRowScrollerR);
-      this.footerRowR = Utils.createDomElement('div', { className: 'slick-footerrow-columns slick-footerrow-columns-right' }, this.footerRowScrollerR);
-      this.footerRow.push(this.footerRowR);
-
-      if (!o.showFooterRow) {
-        Utils.hide(this.footerRowScrollerR);
-      }
-    }
-
-    return true;
-  }
-
-  /**
-   * Builds the right-frozen column band (Phase 4): three panes with NEW
-   * `*-right-frozen` css classes, appended AFTER the six classic panes so classic
-   * sibling positions are untouched, plus header/header-row/top-panel chrome,
-   * viewports and canvases. Shared element arrays are extended at the END so the
-   * classic indexes 0–3 (and [L, R] pairs) stay valid for every existing consumer.
-   * Idempotent: returns false when the band already exists.
-   *
-   * The historical "right" elements keep their class names and become the scrollable
-   * MIDDLE band while this band is active.
-   */
-  materializeRightFrozenBand(o: ViewportMgrBuildOptions): boolean {
-    if (this.paneHeaderRF) {
-      return false;
-    }
-
-    const container = this.container;
-
-    // panes — appended after the classic six (still before the trailing focus sink,
-    // which the grid appends after all panes)
-    this.paneHeaderRF = Utils.createDomElement('div', { className: 'slick-pane slick-pane-header slick-pane-right-frozen', tabIndex: 0 });
-    this.paneTopRF = Utils.createDomElement('div', { className: 'slick-pane slick-pane-top slick-pane-right-frozen', tabIndex: 0 });
-    this.paneBottomRF = Utils.createDomElement('div', { className: 'slick-pane slick-pane-bottom slick-pane-right-frozen', tabIndex: 0 });
-    // insert as a block after the last classic pane (paneBottomR when it exists,
-    // else the lazy grid's paneTopL)
-    const lastClassicPane = this.paneBottomR ?? this.paneTopL;
-    container.insertBefore(this.paneHeaderRF, lastClassicPane.nextSibling);
-    container.insertBefore(this.paneTopRF, this.paneHeaderRF.nextSibling);
-    container.insertBefore(this.paneBottomRF, this.paneTopRF.nextSibling);
-
-    // header chrome
-    this.headerScrollerRF = Utils.createDomElement('div', { className: 'slick-header ui-state-default slick-state-default slick-header-right-frozen' }, this.paneHeaderRF);
-    this.headerScroller.push(this.headerScrollerRF);
-    this.headerRF = Utils.createDomElement('div', { className: 'slick-header-columns slick-header-columns-right-frozen', role: 'row', style: { left: '-1000px' } }, this.headerScrollerRF);
-    this.headers.push(this.headerRF);
-
-    // header row
-    this.headerRowScrollerRF = Utils.createDomElement('div', { className: 'slick-headerrow ui-state-default slick-state-default' }, this.paneTopRF);
-    this.headerRowScroller.push(this.headerRowScrollerRF);
-    this.headerRowSpacerRF = Utils.createDomElement('div', { style: { display: 'block', height: '1px', position: 'absolute', top: '0px', left: '0px' } }, this.headerRowScrollerRF);
-    this.headerRowRF = Utils.createDomElement('div', { className: 'slick-headerrow-columns slick-headerrow-columns-right-frozen' }, this.headerRowScrollerRF);
-    this.headerRows.push(this.headerRowRF);
-
-    // top panel
-    this.topPanelScrollerRF = Utils.createDomElement('div', { className: 'slick-top-panel-scroller ui-state-default slick-state-default' }, this.paneTopRF);
-    this.topPanelScrollers.push(this.topPanelScrollerRF);
-    this.topPanelRF = Utils.createDomElement('div', { className: 'slick-top-panel', style: { width: '10000px' } }, this.topPanelScrollerRF);
-    this.topPanels.push(this.topPanelRF);
-
-    if (!o.showColumnHeader) {
-      Utils.hide(this.headerScrollerRF);
-    }
-    if (!o.showTopPanel) {
-      Utils.hide(this.topPanelScrollerRF);
-    }
-    if (!o.showHeaderRow) {
-      Utils.hide(this.headerRowScrollerRF);
-    }
-
-    // viewports and canvases (array order extended at the END: classic 0–3 preserved)
-    this.viewportTopRF = Utils.createDomElement('div', { className: 'slick-viewport slick-viewport-top slick-viewport-right-frozen', tabIndex: 0 }, this.paneTopRF);
-    this.viewportBottomRF = Utils.createDomElement('div', { className: 'slick-viewport slick-viewport-bottom slick-viewport-right-frozen', tabIndex: 0 }, this.paneBottomRF);
-    this.viewport.push(this.viewportTopRF, this.viewportBottomRF);
-    if (o.viewportClass) {
-      const viewportClassList = Utils.classNameToList(o.viewportClass);
-      this.viewportTopRF.classList.add(...viewportClassList);
-      this.viewportBottomRF.classList.add(...viewportClassList);
-    }
-
-    this.canvasTopRF = Utils.createDomElement('div', { className: 'grid-canvas grid-canvas-top grid-canvas-right-frozen', tabIndex: 0 }, this.viewportTopRF);
-    this.canvasBottomRF = Utils.createDomElement('div', { className: 'grid-canvas grid-canvas-bottom grid-canvas-right-frozen', tabIndex: 0 }, this.viewportBottomRF);
-    this.canvas.push(this.canvasTopRF, this.canvasBottomRF);
-    this.rfTopSlot = this.canvas.indexOf(this.canvasTopRF);
-    this.rfBottomSlot = this.canvas.indexOf(this.canvasBottomRF);
-
-    // footer row
-    if (o.createFooterRow) {
-      this.footerRowScrollerRF = Utils.createDomElement('div', { className: 'slick-footerrow ui-state-default slick-state-default' }, this.paneTopRF);
-      this.footerRowScroller.push(this.footerRowScrollerRF);
-      this.footerRowSpacerRF = Utils.createDomElement('div', { style: { display: 'block', height: '1px', position: 'absolute', top: '0px', left: '0px' } }, this.footerRowScrollerRF);
-      this.footerRowRF = Utils.createDomElement('div', { className: 'slick-footerrow-columns slick-footerrow-columns-right-frozen' }, this.footerRowScrollerRF);
-      this.footerRow.push(this.footerRowRF);
-
-      if (!o.showFooterRow) {
-        Utils.hide(this.footerRowScrollerRF);
-      }
-    }
-
-    // if the bottom-frozen band already exists, add the shared corner pane
-    this.ensureBottomFrozenRightVariant(o);
-
-    return true;
-  }
-
-  /**
-   * Builds the bottom-frozen row band (Phase 4, simultaneous top+bottom mode): one
-   * pane+viewport+canvas per active column band, appended after all existing panes
-   * with `*-bottom-frozen` css classes. Element arrays extend at the END and the
-   * slots are recorded (bfSlotL/R/RF). Idempotent: returns false when the band
-   * already exists. The right-frozen column variant is built only when that band's
-   * DOM exists at call time; materializeRightFrozenBand adds it later otherwise.
-   */
-  materializeBottomFrozenBand(o: ViewportMgrBuildOptions): boolean {
-    if (this.paneBottomFrozenL) {
-      // band exists — but the RF column variant may have arrived after us
-      this.ensureBottomFrozenRightVariant(o);
-      return false;
-    }
-
-    const container = this.container;
-    const lastPane = this.paneBottomRF ?? this.paneBottomR ?? this.paneTopL;
-
-    this.paneBottomFrozenL = Utils.createDomElement('div', { className: 'slick-pane slick-pane-bottom-frozen slick-pane-left', tabIndex: 0 });
-    container.insertBefore(this.paneBottomFrozenL, lastPane.nextSibling);
-    this.paneBottomFrozenR = Utils.createDomElement('div', { className: 'slick-pane slick-pane-bottom-frozen slick-pane-right', tabIndex: 0 });
-    container.insertBefore(this.paneBottomFrozenR, this.paneBottomFrozenL.nextSibling);
-
-    this.viewportBottomFrozenL = Utils.createDomElement('div', { className: 'slick-viewport slick-viewport-bottom-frozen slick-viewport-left', tabIndex: 0 }, this.paneBottomFrozenL);
-    this.viewportBottomFrozenR = Utils.createDomElement('div', { className: 'slick-viewport slick-viewport-bottom-frozen slick-viewport-right', tabIndex: 0 }, this.paneBottomFrozenR);
-    this.canvasBottomFrozenL = Utils.createDomElement('div', { className: 'grid-canvas grid-canvas-bottom-frozen grid-canvas-left', tabIndex: 0 }, this.viewportBottomFrozenL);
-    this.canvasBottomFrozenR = Utils.createDomElement('div', { className: 'grid-canvas grid-canvas-bottom-frozen grid-canvas-right', tabIndex: 0 }, this.viewportBottomFrozenR);
-
-    if (o.viewportClass) {
-      const viewportClassList = Utils.classNameToList(o.viewportClass);
-      this.viewportBottomFrozenL.classList.add(...viewportClassList);
-      this.viewportBottomFrozenR.classList.add(...viewportClassList);
-    }
-
-    this.viewport.push(this.viewportBottomFrozenL, this.viewportBottomFrozenR);
-    this.canvas.push(this.canvasBottomFrozenL, this.canvasBottomFrozenR);
-    this.bfSlotL = this.canvas.indexOf(this.canvasBottomFrozenL);
-    this.bfSlotR = this.canvas.indexOf(this.canvasBottomFrozenR);
-
-    this.ensureBottomFrozenRightVariant(o);
-    return true;
-  }
-
-  /** Adds the bottom-frozen × right-frozen corner pane when both bands exist. */
-  protected ensureBottomFrozenRightVariant(o: ViewportMgrBuildOptions) {
-    if (!this.paneBottomFrozenL || !this.paneHeaderRF || this.paneBottomFrozenRF) {
-      return;
-    }
-    this.paneBottomFrozenRF = Utils.createDomElement('div', { className: 'slick-pane slick-pane-bottom-frozen slick-pane-right-frozen', tabIndex: 0 });
-    this.container.insertBefore(this.paneBottomFrozenRF, this.paneBottomFrozenR.nextSibling);
-    this.viewportBottomFrozenRF = Utils.createDomElement('div', { className: 'slick-viewport slick-viewport-bottom-frozen slick-viewport-right-frozen', tabIndex: 0 }, this.paneBottomFrozenRF);
-    this.canvasBottomFrozenRF = Utils.createDomElement('div', { className: 'grid-canvas grid-canvas-bottom-frozen grid-canvas-right-frozen', tabIndex: 0 }, this.viewportBottomFrozenRF);
-    if (o.viewportClass) {
-      this.viewportBottomFrozenRF.classList.add(...Utils.classNameToList(o.viewportClass));
-    }
-    this.viewport.push(this.viewportBottomFrozenRF);
-    this.canvas.push(this.canvasBottomFrozenRF);
-    this.bfSlotRF = this.canvas.indexOf(this.canvasBottomFrozenRF);
-  }
-
-  /** Whether the simultaneous top+bottom row mode is active AND its DOM exists. */
-  hasBottomFrozenBand(): boolean {
-    return this.bands.frozenTopRows > 0 && this.bands.frozenBottomRows > 0 && !!this.paneBottomFrozenL;
-  }
-
-  /**
-   * Whether the row belongs to the bottom-frozen band. BOUNDED on both sides so the
-   * add-new row (index === dataLength) can never be captured by the band. The same
-   * test is used on both the render and lookup sides — the new band deliberately
-   * avoids the historical one-row threshold asymmetry of the legacy single band.
-   */
-  isRowInBottomFrozenBand(row: number): boolean {
-    if (!this.hasBottomFrozenBand()) {
-      return false;
-    }
-    const split = this.freeze.bottomFrozenSplitRow ?? Number.MAX_SAFE_INTEGER;
-    return row >= split && row < split + this.bands.frozenBottomRows;
-  }
-
-  //////////////////////////////////////////////////////////////////////////////////////////////
-  // Freeze state and pane selection (Phase 2 of the encapsulation refactor)
-  //////////////////////////////////////////////////////////////////////////////////////////////
-
-  protected freeze: ViewportFreezeState = { frozenColumnIdx: -1, hasFrozenRows: false, actualFrozenRow: -1, frozenBottom: false };
-  protected bands: FreezeBandCounts = { frozenLeftCols: 0, frozenRightCols: 0, frozenTopRows: 0, frozenBottomRows: 0 };
-
-  /** Receives the grid's freeze configuration; called by SlickGrid.setFrozenOptions(). */
-  updateFreezeState(f: ViewportFreezeState) {
-    this.freeze = { ...f };
-
-    // derive the band-count view (Phase 4 groundwork); the legacy fields above stay
-    // authoritative for the existing 2×2 code paths
-    const rowCount = f.hasFrozenRows ? Math.max(0, f.frozenRowCount ?? 0) : 0;
-    const bottomRowCount = Math.max(0, f.frozenBottomRowCount ?? 0);
-    this.bands = {
-      frozenLeftCols: f.frozenColumnIdx + 1,
-      frozenRightCols: Math.max(0, f.frozenRightColCount ?? 0),
-      // with an explicit bottom count, frozenRow always means TOP rows and the legacy
-      // frozenBottom flag is ignored (it only positions the single-band case)
-      frozenTopRows: bottomRowCount > 0 ? rowCount : (f.frozenBottom ? 0 : rowCount),
-      frozenBottomRows: bottomRowCount > 0 ? bottomRowCount : (f.frozenBottom ? rowCount : 0),
-    };
-  }
-
-  /** Band-count view of the freeze configuration (zero count = band does not exist). */
-  bandCounts(): FreezeBandCounts {
-    return this.bands;
-  }
-
-  /** Returns a boolean indicating whether the grid is configured with frozen columns. */
-  hasFrozenColumns() {
-    return this.bands.frozenLeftCols > 0;
-  }
-
-  /** Returns a boolean indicating whether the grid is configured with frozen rows. */
-  hasFrozenRows() {
-    return this.freeze.hasFrozenRows;
-  }
-
-  /**
-   * The left canvas of the scrollable body band: bottom-left while rows are frozen at
-   * the top, top-left otherwise (historical selector used by updateRowCount and
-   * bindAncestorScrollEvents).
-   */
-  bodyCanvasL(): HTMLDivElement {
-    return (this.freeze.hasFrozenRows && !this.freeze.frozenBottom) ? this.canvasBottomL : this.canvasTopL;
-  }
-
-  /**
-   * Index of the pane owning cell (colIdx, rowIdx) in the element arrays:
-   * classic slots [TopL, TopR, BottomL, BottomR], right-frozen slots [TopRF, BottomRF]
-   * appended at 4/5 (materializeRightFrozenPanes canonicalizes the classic set first,
-   * so these positions hold under lazyPanes too).
-   */
-  paneCellIndex(colIdx: number, rowIdx: number): number {
-    if (this.isRowInBottomFrozenBand(rowIdx)) {
-      if (this.isColumnInRightFrozenBand(colIdx)) {
-        return this.bfSlotRF;
-      }
-      const isRightSideBF = this.hasFrozenColumns() && colIdx > this.freeze.frozenColumnIdx;
-      return isRightSideBF ? this.bfSlotR : this.bfSlotL;
-    }
-    const isBottomSide = this.freeze.hasFrozenRows && rowIdx >= this.freeze.actualFrozenRow + (this.freeze.frozenBottom ? 0 : 1);
-    if (this.isColumnInRightFrozenBand(colIdx)) {
-      return isBottomSide ? this.rfBottomSlot : this.rfTopSlot;
-    }
-    const isRightSide = this.hasFrozenColumns() && colIdx > this.freeze.frozenColumnIdx;
-    return (isBottomSide ? 2 : 0) + (isRightSide ? 1 : 0);
-  }
-
-  /**
-   * Get frozen (pinned) row offset
-   *
-   * Returns the vertical pixel offset to apply for frozen rows.
-   * Depending on whether frozen rows are pinned at the bottom or top and based on grid height,
-   * it returns either a fixed frozen rows height or a calculated offset.
-   *
-   * @param {Number} row - grid row number
-   */
-  frozenRowOffset(row: number, g: { h: number; viewportTopH: number; frozenRowsHeight: number; rowHeight: number; }): number {
-    // bottom-frozen band (simultaneous mode): rebase to band-local coordinates
-    if (this.isRowInBottomFrozenBand(row)) {
-      return this.freeze.bottomFrozenSplitRow! * g.rowHeight;
-    }
-
-    // let offset = ( hasFrozenRows ) ? ( this._options.frozenBottom ) ? ( row >= actualFrozenRow ) ? ( h < viewportTopH ) ? ( actualFrozenRow * this._options.rowHeight ) : h : 0 : ( row >= actualFrozenRow ) ? frozenRowsHeight : 0 : 0; // WTF?
-    let offset = 0;
-    if (this.freeze.hasFrozenRows) {
-      if (this.freeze.frozenBottom) {
-        if (row >= this.freeze.actualFrozenRow) {
-          if (g.h < g.viewportTopH) {
-            offset = (this.freeze.actualFrozenRow * g.rowHeight);
-          } else {
-            offset = g.h;
-          }
-        } else {
-          offset = 0;
-        }
-      }
-      else {
-        if (row >= this.freeze.actualFrozenRow) {
-          offset = g.frozenRowsHeight;
-        } else {
-          offset = 0;
-        }
-      }
-    } else {
-      offset = 0;
-    }
-
-    return offset;
-  }
-
-  /**
-   * Whether the row lives in a frozen band and must therefore be kept out of row
-   * virtualization cleanup (historical cleanupRows predicate).
-   */
-  isRowInFrozenBand(row: number): boolean {
-    if (this.isRowInBottomFrozenBand(row)) {
-      return true;
-    }
-    return this.freeze.hasFrozenRows
-      && ((this.freeze.frozenBottom && row >= this.freeze.actualFrozenRow) // Frozen bottom rows
-        || (!this.freeze.frozenBottom && row <= this.freeze.actualFrozenRow) // Frozen top rows
-      );
-  }
-
-  /**
-   * Whether cell-level cleanup must skip the row entirely (historical cleanUpCells
-   * predicate). NOTE: transcribed verbatim — the second disjunct is NOT guarded by
-   * !frozenBottom, so for frozenBottom grids every row is exempt; that quirk is
-   * long-standing upstream behaviour and is deliberately preserved.
-   */
-  isRowCellCleanupExempt(row: number): boolean {
-    if (this.isRowInBottomFrozenBand(row)) {
-      return true;
-    }
-    return this.freeze.hasFrozenRows
-      && ((this.freeze.frozenBottom && row > this.freeze.actualFrozenRow) // Frozen bottom rows
-        || (row <= this.freeze.actualFrozenRow)                     // Frozen top rows
-      );
-  }
-
-  /** Whether the column index falls inside the left frozen band. */
-  isColumnInFrozenBand(colIdx: number): boolean {
-    return colIdx <= this.freeze.frozenColumnIdx;
-  }
-
-  /** True when frozen columns are on AND the column index falls right of the freeze. */
-  isColumnRightOfFreeze(colIdx: number): boolean {
-    return this.hasFrozenColumns() && colIdx > this.freeze.frozenColumnIdx;
-  }
-
-  /** Column index local to its side container (right-side children are indexed after the freeze). */
-  sideLocalColumnIdx(colIdx: number): number {
-    return this.isColumnRightOfFreeze(colIdx) ? colIdx - this.freeze.frozenColumnIdx - 1 : colIdx;
-  }
-
-  /** Pick the left or right element of an [L, R] pair for the given column. */
-  sideForColumn<T>(colIdx: number, left: T, right: T): T {
-    return this.isColumnRightOfFreeze(colIdx) ? right : left;
-  }
-
-  /** Whether the right-frozen band is active AND its DOM has been materialized. */
-  hasRightFrozenBand(): boolean {
-    return this.bands.frozenRightCols > 0 && !!this.paneHeaderRF;
-  }
-
-  /** Whether the column index falls inside the right-frozen band. */
-  isColumnInRightFrozenBand(colIdx: number): boolean {
-    return this.bands.frozenRightCols > 0 && colIdx >= (this.freeze.frozenRightStartIdx ?? Number.MAX_SAFE_INTEGER);
-  }
-
-  /** Three-way band pick: left band, scrollable middle, or right-frozen element. */
-  bandElementForColumn<T>(colIdx: number, left: T, right: T, rightFrozen: T): T {
-    if (this.isColumnInRightFrozenBand(colIdx)) {
-      return rightFrozen;
-    }
-    return this.sideForColumn(colIdx, left, right);
-  }
-
-  /** Column index local to its band container (right-frozen children index from the band start). */
-  bandLocalColumnIdx(colIdx: number): number {
-    if (this.isColumnInRightFrozenBand(colIdx)) {
-      return colIdx - this.freeze.frozenRightStartIdx!;
-    }
-    return this.sideLocalColumnIdx(colIdx);
-  }
-
-  /**
-   * Index of the column's node inside rowsCache[].rowNode: 0 for the left fragment
-   * (which is the scrollable fragment when no columns are left-frozen), 1 for the
-   * middle fragment under a left freeze, and last for the right-frozen fragment.
-   */
-  rowNodeIdxForColumn(colIdx: number): number {
-    if (this.isColumnInRightFrozenBand(colIdx)) {
-      return this.hasFrozenColumns() ? 2 : 1;
-    }
-    return this.isColumnRightOfFreeze(colIdx) ? 1 : 0;
-  }
-
-  /** Utils.show that tolerates panes not built under lazyPanes. */
-  protected showIf(el?: HTMLElement) {
-    if (el) { Utils.show(el); }
-  }
-
-  /** Utils.hide that tolerates panes not built under lazyPanes. */
-  protected hideIf(el?: HTMLElement) {
-    if (el) { Utils.hide(el); }
-  }
-
-  /** add/remove frozen class to left headers/footer when defined */
-  applyPaneFrozenClasses(): void {
-    const classAction = this.hasFrozenColumns() ? 'add' : 'remove';
-    for (const elm of [this.paneHeaderL, this.paneTopL, this.paneBottomL]) {
-      elm?.classList[classAction]('frozen');
-    }
-  }
-
-  /** Shows/hides the right and bottom panes according to the freeze configuration. */
-  applyPaneVisibility() {
-    if (this.hasFrozenColumns()) {
-      this.showIf(this.paneHeaderR);
-      this.showIf(this.paneTopR);
-
-      if (this.freeze.hasFrozenRows) {
-        this.showIf(this.paneBottomL);
-        this.showIf(this.paneBottomR);
-      } else {
-        this.hideIf(this.paneBottomR);
-        this.hideIf(this.paneBottomL);
-      }
-    } else {
-      this.hideIf(this.paneHeaderR);
-      this.hideIf(this.paneTopR);
-      this.hideIf(this.paneBottomR);
-
-      if (this.freeze.hasFrozenRows) {
-        this.showIf(this.paneBottomL);
-      } else {
-        this.hideIf(this.paneBottomR);
-        this.hideIf(this.paneBottomL);
-      }
-    }
-
-    // right-frozen band (exists only after materialization; kept hidden — like the
-    // classic panes — when the right freeze is turned off again)
-    if (this.bands.frozenRightCols > 0) {
-      this.showIf(this.paneHeaderRF);
-      this.showIf(this.paneTopRF);
-      if (this.freeze.hasFrozenRows) {
-        this.showIf(this.paneBottomRF);
-      } else {
-        this.hideIf(this.paneBottomRF);
-      }
-    } else {
-      this.hideIf(this.paneHeaderRF);
-      this.hideIf(this.paneTopRF);
-      this.hideIf(this.paneBottomRF);
-    }
-
-    // bottom-frozen row band (simultaneous top+bottom mode only); column-band
-    // visibility mirrors the classic bottom panes
-    if (this.bands.frozenTopRows > 0 && this.bands.frozenBottomRows > 0) {
-      this.showIf(this.paneBottomFrozenL);
-      if (this.hasFrozenColumns()) {
-        this.showIf(this.paneBottomFrozenR);
-      } else {
-        this.hideIf(this.paneBottomFrozenR);
-      }
-      if (this.bands.frozenRightCols > 0) {
-        this.showIf(this.paneBottomFrozenRF);
-      } else {
-        this.hideIf(this.paneBottomFrozenRF);
-      }
-    } else {
-      this.hideIf(this.paneBottomFrozenL);
-      this.hideIf(this.paneBottomFrozenR);
-      this.hideIf(this.paneBottomFrozenRF);
-    }
-  }
-
-  /**
-   * Sets the CSS overflowX and overflowY styles for all four viewport elements
-   * (top–left, top–right, bottom–left, bottom–right) based on the freeze configuration
-   * and options such as alwaysAllowHorizontalScroll and alwaysShowVerticalScroll.
-   * If a viewportClass is specified in options, the class is added to each viewport.
-   */
-  applyOverflow(o: { alwaysAllowHorizontalScroll?: boolean; alwaysShowVerticalScroll?: boolean; viewportClass?: string; }) {
-    const hasFrozenRows = this.freeze.hasFrozenRows;
-    this.viewportTopL.style.overflowX = (this.hasFrozenColumns()) ? (hasFrozenRows && !o.alwaysAllowHorizontalScroll ? 'hidden' : 'scroll') : (hasFrozenRows && !o.alwaysAllowHorizontalScroll ? 'hidden' : 'auto');
-    this.viewportTopL.style.overflowY = (!this.hasFrozenColumns() && o.alwaysShowVerticalScroll) ? 'scroll' : ((this.hasFrozenColumns()) ? (hasFrozenRows ? 'hidden' : 'hidden') : (hasFrozenRows ? 'scroll' : 'auto'));
-
-    if (this.viewportTopR) {
-      this.viewportTopR.style.overflowX = (this.hasFrozenColumns()) ? (hasFrozenRows && !o.alwaysAllowHorizontalScroll ? 'hidden' : 'scroll') : (hasFrozenRows && !o.alwaysAllowHorizontalScroll ? 'hidden' : 'auto');
-      this.viewportTopR.style.overflowY = o.alwaysShowVerticalScroll ? 'scroll' : ((this.hasFrozenColumns()) ? (hasFrozenRows ? 'scroll' : 'auto') : (hasFrozenRows ? 'scroll' : 'auto'));
-    }
-
-    if (this.viewportBottomL) {
-      this.viewportBottomL.style.overflowX = (this.hasFrozenColumns()) ? (hasFrozenRows && !o.alwaysAllowHorizontalScroll ? 'scroll' : 'auto') : (hasFrozenRows && !o.alwaysAllowHorizontalScroll ? 'auto' : 'auto');
-      this.viewportBottomL.style.overflowY = (!this.hasFrozenColumns() && o.alwaysShowVerticalScroll) ? 'scroll' : ((this.hasFrozenColumns()) ? (hasFrozenRows ? 'hidden' : 'hidden') : (hasFrozenRows ? 'scroll' : 'auto'));
-    }
-
-    if (this.viewportBottomR) {
-      this.viewportBottomR.style.overflowX = (this.hasFrozenColumns()) ? (hasFrozenRows && !o.alwaysAllowHorizontalScroll ? 'scroll' : 'auto') : (hasFrozenRows && !o.alwaysAllowHorizontalScroll ? 'auto' : 'auto');
-      this.viewportBottomR.style.overflowY = o.alwaysShowVerticalScroll ? 'scroll' : ((this.hasFrozenColumns()) ? (hasFrozenRows ? 'auto' : 'auto') : (hasFrozenRows ? 'auto' : 'auto'));
-    }
-
-    // bottom-frozen viewports never own a scrollbar: Y is fixed, X follows the
-    // scroll owner programmatically
-    if (this.viewportBottomFrozenL) {
-      this.viewportBottomFrozenL.style.overflowX = 'hidden';
-      this.viewportBottomFrozenL.style.overflowY = 'hidden';
-    }
-    if (this.viewportBottomFrozenR) {
-      this.viewportBottomFrozenR.style.overflowX = 'hidden';
-      this.viewportBottomFrozenR.style.overflowY = 'hidden';
-    }
-    if (this.viewportBottomFrozenRF) {
-      this.viewportBottomFrozenRF.style.overflowX = 'hidden';
-      this.viewportBottomFrozenRF.style.overflowY = 'hidden';
-    }
-
-    // right-frozen viewports never own a scrollbar: X is fixed, Y follows the
-    // scroll owner programmatically (same rationale as the frozen-left viewport)
-    if (this.viewportTopRF) {
-      this.viewportTopRF.style.overflowX = 'hidden';
-      this.viewportTopRF.style.overflowY = 'hidden';
-    }
-    if (this.viewportBottomRF) {
-      this.viewportBottomRF.style.overflowX = 'hidden';
-      this.viewportBottomRF.style.overflowY = 'hidden';
-    }
-
-    if (o.viewportClass) {
-      const viewportClassList = Utils.classNameToList(o.viewportClass);
-      // this.viewport only ever contains the elements that were actually built
-      this.viewport.forEach((view) => {
-        view.classList.add(...viewportClassList);
-      });
-    }
-  }
-
-  /**
-   * Picks which viewport owns the X and Y scrollbars and which header/header-row/footer-row
-   * scrollers follow horizontal scrolling, according to the freeze configuration.
-   * The horizontal scrollbar must sit at the physical bottom of the grid, which is why
-   * frozenBottom splits X and Y ownership.
-   */
-  /**
-   * Distributes computed canvas/header widths onto the pane, viewport, canvas, header,
-   * header-row and footer-row elements. Transcribed from the historical middle section
-   * of SlickGrid.updateCanvasWidth(); the width computations stay in the grid.
-   */
-  applyCanvasWidths(g: CanvasWidthsGeometry) {
-    // width reserved by the right-frozen band (0 while the band is off or not built);
-    // the scrollable middle band shrinks by this amount
-    const rfActive = this.bands.frozenRightCols > 0 && !!this.paneHeaderRF;
-    const rfW = rfActive ? g.canvasWidthRF : 0;
-
-    if (g.widthChanged || this.hasFrozenColumns() || this.freeze.hasFrozenRows || rfActive) {
-      Utils.width(this.canvasTopL, g.canvasWidthL);
-
-      Utils.width(this.headerL, g.headersWidthL);
-      if (this.headerR) {
-        Utils.width(this.headerR, g.headersWidthR);
-      }
-
-      if (this.hasFrozenColumns()) {
-        Utils.width(this.canvasTopR, g.canvasWidthR);
-
-        Utils.width(this.paneHeaderL, g.canvasWidthL);
-        Utils.setStyleSize(this.paneHeaderR, 'left', g.canvasWidthL);
-        Utils.setStyleSize(this.paneHeaderR, 'width', g.viewportW - g.canvasWidthL - rfW);
-
-        Utils.width(this.paneTopL, g.canvasWidthL);
-        Utils.setStyleSize(this.paneTopR, 'left', g.canvasWidthL);
-        Utils.width(this.paneTopR, g.viewportW - g.canvasWidthL - rfW);
-
-        Utils.width(this.headerRowScrollerL, g.canvasWidthL);
-        Utils.width(this.headerRowScrollerR, g.viewportW - g.canvasWidthL - rfW);
-
-        Utils.width(this.headerRowL, g.canvasWidthL);
-        Utils.width(this.headerRowR, g.canvasWidthR);
-
-        if (g.createFooterRow) {
-          Utils.width(this.footerRowScrollerL, g.canvasWidthL);
-          Utils.width(this.footerRowScrollerR, g.viewportW - g.canvasWidthL - rfW);
-
-          Utils.width(this.footerRowL, g.canvasWidthL);
-          Utils.width(this.footerRowR, g.canvasWidthR);
-        }
-        if (g.createPreHeaderPanel) {
-          Utils.width(this.preHeaderPanel, g.preHeaderPanelWidth ?? g.canvasWidth);
-        }
-        Utils.width(this.viewportTopL, g.canvasWidthL);
-        Utils.width(this.viewportTopR, g.viewportW - g.canvasWidthL - rfW);
-
-        if (this.freeze.hasFrozenRows) {
-          Utils.width(this.paneBottomL, g.canvasWidthL);
-          Utils.setStyleSize(this.paneBottomR, 'left', g.canvasWidthL);
-
-          Utils.width(this.viewportBottomL, g.canvasWidthL);
-          Utils.width(this.viewportBottomR, g.viewportW - g.canvasWidthL - rfW);
-
-          Utils.width(this.canvasBottomL, g.canvasWidthL);
-          Utils.width(this.canvasBottomR, g.canvasWidthR);
-        }
-      } else if (rfActive) {
-        // no left freeze, but a right-frozen band: the left pane IS the scrollable
-        // middle band — pixel widths instead of the historical '100%'
-        const middleW = g.viewportW - rfW;
-        Utils.width(this.paneHeaderL, middleW);
-        Utils.width(this.paneTopL, middleW);
-        Utils.width(this.headerRowScrollerL, middleW);
-        Utils.width(this.headerRowL, g.canvasWidth);
-
-        if (g.createFooterRow) {
-          Utils.width(this.footerRowScrollerL, middleW);
-          Utils.width(this.footerRowL, g.canvasWidth);
-        }
-
-        if (g.createPreHeaderPanel) {
-          Utils.width(this.preHeaderPanel, g.preHeaderPanelWidth ?? g.canvasWidth);
-        }
-        Utils.width(this.viewportTopL, middleW);
-
-        if (this.freeze.hasFrozenRows) {
-          Utils.width(this.viewportBottomL, middleW);
-          Utils.width(this.canvasBottomL, g.canvasWidthL);
-        }
-      } else {
-        Utils.width(this.paneHeaderL, '100%');
-        Utils.width(this.paneTopL, '100%');
-        Utils.width(this.headerRowScrollerL, '100%');
-        Utils.width(this.headerRowL, g.canvasWidth);
-
-        if (g.createFooterRow) {
-          Utils.width(this.footerRowScrollerL, '100%');
-          Utils.width(this.footerRowL, g.canvasWidth);
-        }
-
-        if (g.createPreHeaderPanel) {
-          Utils.width(this.preHeaderPanel, g.preHeaderPanelWidth ?? g.canvasWidth);
-        }
-        Utils.width(this.viewportTopL, '100%');
-
-        if (this.freeze.hasFrozenRows) {
-          Utils.width(this.viewportBottomL, '100%');
-          Utils.width(this.canvasBottomL, g.canvasWidthL);
-        }
-      }
-
-      // bottom-frozen row band (simultaneous mode): column widths mirror the classic
-      // bottom panes
-      if (this.hasBottomFrozenBand()) {
-        if (this.hasFrozenColumns()) {
-          Utils.width(this.paneBottomFrozenL, g.canvasWidthL);
-          Utils.width(this.canvasBottomFrozenL, g.canvasWidthL);
-          Utils.setStyleSize(this.paneBottomFrozenR, 'left', g.canvasWidthL);
-          Utils.width(this.paneBottomFrozenR, g.viewportW - g.canvasWidthL - rfW);
-          Utils.width(this.viewportBottomFrozenR, g.viewportW - g.canvasWidthL - rfW);
-          Utils.width(this.canvasBottomFrozenR, g.canvasWidthR);
-          Utils.width(this.viewportBottomFrozenL, g.canvasWidthL);
-        } else if (rfActive) {
-          Utils.width(this.paneBottomFrozenL, g.viewportW - rfW);
-          Utils.width(this.viewportBottomFrozenL, g.viewportW - rfW);
-          Utils.width(this.canvasBottomFrozenL, g.canvasWidthL);
-        } else {
-          Utils.width(this.paneBottomFrozenL, '100%');
-          Utils.width(this.viewportBottomFrozenL, '100%');
-          Utils.width(this.canvasBottomFrozenL, g.canvasWidthL);
-        }
-
-        if (this.paneBottomFrozenRF && rfActive) {
-          Utils.setStyleSize(this.paneBottomFrozenRF, 'left', g.viewportW - rfW);
-          Utils.width(this.paneBottomFrozenRF, rfW);
-          Utils.width(this.viewportBottomFrozenRF, rfW);
-          Utils.width(this.canvasBottomFrozenRF, g.canvasWidthRF);
-        }
-      }
-
-      // right-frozen band: fixed-width panes pinned to the right edge
-      if (rfActive) {
-        const rfLeft = g.viewportW - rfW;
-        Utils.setStyleSize(this.paneHeaderRF, 'left', rfLeft);
-        Utils.width(this.paneHeaderRF, rfW);
-        Utils.width(this.headerRF, g.headersWidthRF);
-
-        Utils.setStyleSize(this.paneTopRF, 'left', rfLeft);
-        Utils.width(this.paneTopRF, rfW);
-        Utils.width(this.headerRowScrollerRF, rfW);
-        Utils.width(this.headerRowRF, g.canvasWidthRF);
-        Utils.width(this.viewportTopRF, rfW);
-        Utils.width(this.canvasTopRF, g.canvasWidthRF);
-
-        if (g.createFooterRow) {
-          Utils.width(this.footerRowScrollerRF, rfW);
-          Utils.width(this.footerRowRF, g.canvasWidthRF);
-        }
-
-        if (this.freeze.hasFrozenRows) {
-          Utils.setStyleSize(this.paneBottomRF, 'left', rfLeft);
-          Utils.width(this.paneBottomRF, rfW);
-          Utils.width(this.viewportBottomRF, rfW);
-          Utils.width(this.canvasBottomRF, g.canvasWidthRF);
-        }
-      }
-    }
-
-    Utils.width(this.headerRowSpacerL, g.canvasWidth + (g.viewportHasVScroll ? g.scrollbarWidth : 0));
-    if (this.headerRowSpacerR) {
-      Utils.width(this.headerRowSpacerR, g.canvasWidth + (g.viewportHasVScroll ? g.scrollbarWidth : 0));
-    }
-
-    if (g.createFooterRow) {
-      Utils.width(this.footerRowSpacerL, g.canvasWidth + (g.viewportHasVScroll ? g.scrollbarWidth : 0));
-      if (this.footerRowSpacerR) {
-        Utils.width(this.footerRowSpacerR, g.canvasWidth + (g.viewportHasVScroll ? g.scrollbarWidth : 0));
-      }
-    }
-  }
-
-  /**
-   * Computes the pane/viewport heights from the freeze configuration and distributes them
-   * onto the pane, viewport and canvas elements. Transcribed from the historical middle
-   * section of SlickGrid.resizeCanvas(); returns the computed heights for the grid to keep.
-   */
-  applyPaneHeights(g: PaneHeightsGeometry): { paneTopH: number; paneBottomH: number; viewportTopH: number; viewportBottomH: number; } {
-    let paneTopH = 0;
-    let paneBottomH = 0;
-    let viewportTopH = 0;
-    const viewportBottomH = 0;
-
-    // Account for Frozen Rows
-    const simultaneousBands = this.bands.frozenTopRows > 0 && this.bands.frozenBottomRows > 0 && !!this.paneBottomFrozenL;
-    if (this.freeze.hasFrozenRows) {
-      if (this.freeze.frozenBottom) {
-        paneTopH = g.viewportH - g.frozenRowsHeight - g.scrollbarHeight;
-        paneBottomH = g.frozenRowsHeight + g.scrollbarHeight;
-      } else {
-        paneTopH = g.frozenRowsHeight;
-        paneBottomH = g.viewportH - g.frozenRowsHeight;
-        if (simultaneousBands) {
-          // the scrollable body shrinks by the bottom-frozen band height
-          paneBottomH -= g.frozenBottomRowsHeight ?? 0;
-        }
-      }
-    } else {
-      paneTopH = g.viewportH;
-    }
-
-    // The top pane includes the top panel and the header row
-    paneTopH += g.topPanelH + g.headerRowH + g.footerRowH;
-
-    if (this.hasFrozenColumns() && g.autoHeight) {
-      paneTopH += g.scrollbarHeight;
-    }
-
-    // The top viewport does not contain the top panel or header row
-    viewportTopH = paneTopH - g.topPanelH - g.headerRowH - g.footerRowH;
-
-    if (g.autoHeight) {
-      if (this.hasFrozenColumns()) {
-        let fullHeight = paneTopH + this.headerScrollerL.offsetHeight;
-        fullHeight += g.getContainerVBoxDelta();
-        if (g.showPreHeaderPanel) {
-          fullHeight += g.preHeaderPanelHeight!;
-        }
-        Utils.height(this.container, fullHeight);
-      }
-
-      this.paneTopL.style.position = 'relative';
-    }
-
-    let topHeightOffset = Utils.height(this.paneHeaderL);
-    if (topHeightOffset) {
-      topHeightOffset += (g.showTopHeaderPanel ? g.topHeaderPanelHeight! : 0);
-    } else {
-      topHeightOffset = (g.showHeaderRow ? g.headerRowHeight! : 0) + (g.showPreHeaderPanel ? g.preHeaderPanelHeight! : 0);
-    }
-    Utils.setStyleSize(this.paneTopL, 'top', topHeightOffset || topHeightOffset);
-    Utils.height(this.paneTopL, paneTopH);
-
-    const paneBottomTop = this.paneTopL.offsetTop + paneTopH;
-
-    if (!g.autoHeight) {
-      Utils.height(this.viewportTopL, viewportTopH);
-    }
-
-    if (this.hasFrozenColumns()) {
-      let topHeightOffset = Utils.height(this.paneHeaderL);
-      if (topHeightOffset) {
-        topHeightOffset += (g.showTopHeaderPanel ? g.topHeaderPanelHeight! : 0);
-      }
-      Utils.setStyleSize(this.paneTopR, 'top', topHeightOffset as number);
-      Utils.height(this.paneTopR, paneTopH);
-      Utils.height(this.viewportTopR, viewportTopH);
-
-      if (this.freeze.hasFrozenRows) {
-        Utils.setStyleSize(this.paneBottomL, 'top', paneBottomTop);
-        Utils.height(this.paneBottomL, paneBottomH);
-        Utils.setStyleSize(this.paneBottomR, 'top', paneBottomTop);
-        Utils.height(this.paneBottomR, paneBottomH);
-        Utils.height(this.viewportBottomR, paneBottomH);
-      }
-    } else {
-      if (this.freeze.hasFrozenRows) {
-        Utils.width(this.paneBottomL, '100%');
-        Utils.height(this.paneBottomL, paneBottomH);
-        Utils.setStyleSize(this.paneBottomL, 'top', paneBottomTop);
-      }
-    }
-
-    if (this.freeze.hasFrozenRows) {
-      Utils.height(this.viewportBottomL, paneBottomH);
-
-      if (this.freeze.frozenBottom) {
-        Utils.height(this.canvasBottomL, g.frozenRowsHeight);
-
-        if (this.hasFrozenColumns()) {
-          Utils.height(this.canvasBottomR, g.frozenRowsHeight);
-        }
-      } else {
-        Utils.height(this.canvasTopL, g.frozenRowsHeight);
-
-        if (this.hasFrozenColumns()) {
-          Utils.height(this.canvasTopR, g.frozenRowsHeight);
-        }
-      }
-    } else {
-      if (this.viewportTopR) {
-        Utils.height(this.viewportTopR, viewportTopH);
-      }
-    }
-
-    // bottom-frozen row band (simultaneous mode): pinned below the shrunk body pane
-    if (simultaneousBands) {
-      const bfH = g.frozenBottomRowsHeight ?? 0;
-      const bfTop = this.paneTopL.offsetTop + paneTopH + paneBottomH;
-
-      Utils.setStyleSize(this.paneBottomFrozenL, 'top', bfTop);
-      Utils.height(this.paneBottomFrozenL, bfH);
-      Utils.height(this.viewportBottomFrozenL, bfH);
-      Utils.height(this.canvasBottomFrozenL, bfH);
-
-      if (this.hasFrozenColumns()) {
-        Utils.setStyleSize(this.paneBottomFrozenR, 'top', bfTop);
-        Utils.height(this.paneBottomFrozenR, bfH);
-        Utils.height(this.viewportBottomFrozenR, bfH);
-        Utils.height(this.canvasBottomFrozenR, bfH);
-      }
-
-      if (this.paneBottomFrozenRF) {
-        Utils.setStyleSize(this.paneBottomFrozenRF, 'top', bfTop);
-        Utils.height(this.paneBottomFrozenRF, bfH);
-        Utils.height(this.viewportBottomFrozenRF, bfH);
-        Utils.height(this.canvasBottomFrozenRF, bfH);
-      }
-    }
-
-    // right-frozen band: mirror the classic right-pane vertical geometry
-    if (this.bands.frozenRightCols > 0 && this.paneHeaderRF) {
-      let topHeightOffsetRF = Utils.height(this.paneHeaderL);
-      if (topHeightOffsetRF) {
-        topHeightOffsetRF += (g.showTopHeaderPanel ? g.topHeaderPanelHeight! : 0);
-      }
-      Utils.setStyleSize(this.paneTopRF, 'top', topHeightOffsetRF as number);
-      Utils.height(this.paneTopRF, paneTopH);
-      Utils.height(this.viewportTopRF, viewportTopH);
-
-      if (this.freeze.hasFrozenRows) {
-        const paneBottomTopRF = this.paneTopL.offsetTop + paneTopH;
-        Utils.setStyleSize(this.paneBottomRF, 'top', paneBottomTopRF);
-        Utils.height(this.paneBottomRF, paneBottomH);
-        Utils.height(this.viewportBottomRF, paneBottomH);
-
-        if (this.freeze.frozenBottom) {
-          Utils.height(this.canvasBottomRF, g.frozenRowsHeight);
-        } else {
-          Utils.height(this.canvasTopRF, g.frozenRowsHeight);
-        }
-      }
-    }
-
-    return { paneTopH, paneBottomH, viewportTopH, viewportBottomH };
-  }
-
-  /** the current scroll-owner/follower set, refreshed by selectScrollContainers() */
-  protected scrollContainers!: { x: HTMLDivElement; y: HTMLDivElement; header: HTMLDivElement; headerRow: HTMLDivElement; footerRow: HTMLDivElement; };
-
-  /**
-   * Attaches one rendered row (left fragment + right fragment when columns are frozen)
-   * to the canvases owned by the row's band, returning the rowNode array for the grid's
-   * rowsCache (or null if the expected fragments are missing).
-   *
-   * NOTE: the band threshold here is `rowIdx >= actualFrozenRow` — deliberately WITHOUT
-   * the `+ (frozenBottom ? 0 : 1)` adjustment used by paneCellIndex(); the historical
-   * render-side and cell-lookup-side splits differ by one row in the non-frozenBottom
-   * case, and that asymmetry is preserved verbatim.
-   */
-  attachRow(rowIdx: number, left: HTMLElement | null, right: HTMLElement | null, rightFrozen?: HTMLElement | null): HTMLElement[] | null {
-    let attached: HTMLElement[] | null = null;
-    const isBFBand = this.isRowInBottomFrozenBand(rowIdx);
-    const isBottomBand = !isBFBand && (this.freeze.hasFrozenRows) && (rowIdx >= this.freeze.actualFrozenRow);
-
-    if (isBFBand) {
-      if (this.hasFrozenColumns()) {
-        if (left && right) {
-          this.canvasBottomFrozenL.appendChild(left);
-          this.canvasBottomFrozenR.appendChild(right);
-          attached = [left, right];
-        }
-      } else if (left) {
-        this.canvasBottomFrozenL.appendChild(left);
-        attached = [left];
-      }
-    } else if (isBottomBand) {
-      if (this.hasFrozenColumns()) {
-        if (left && right) {
-          this.canvasBottomL.appendChild(left);
-          this.canvasBottomR.appendChild(right);
-          attached = [left, right];
-        }
-      } else if (left) {
-        this.canvasBottomL.appendChild(left);
-        attached = [left];
-      }
-    } else if (this.hasFrozenColumns()) {
-      if (left && right) {
-        this.canvasTopL.appendChild(left);
-        this.canvasTopR.appendChild(right);
-        attached = [left, right];
-      }
-    } else if (left) {
-      this.canvasTopL.appendChild(left);
-      attached = [left];
-    }
-
-    // right-frozen fragment always sits LAST in the rowNode array
-    const rfTargetCanvas = isBFBand ? this.canvasBottomFrozenRF : (isBottomBand ? this.canvasBottomRF : this.canvasTopRF);
-    if (attached && this.bands.frozenRightCols > 0 && rfTargetCanvas && rightFrozen) {
-      rfTargetCanvas.appendChild(rightFrozen);
-      attached.push(rightFrozen);
-    }
-
-    return attached;
-  }
-
-  /** Applies an X scroll position to the scroll-owner viewport and every horizontal follower. */
-  syncHorizontalScroll(x: number, o: { createFooterRow?: boolean; createPreHeaderPanel?: boolean; }) {
-    this.scrollContainers.x.scrollLeft = x;
-    this.scrollContainers.header.scrollLeft = x;
-    this.topPanelScrollers[0].scrollLeft = x;
-    if (o.createFooterRow) {
-      this.scrollContainers.footerRow.scrollLeft = x;
-    }
-    if (o.createPreHeaderPanel) {
-      if (this.hasFrozenColumns()) {
-        this.preHeaderPanelScrollerR.scrollLeft = x;
-      } else {
-        this.preHeaderPanelScroller.scrollLeft = x;
-      }
-    }
-
-    if (this.hasFrozenColumns()) {
-      if (this.freeze.hasFrozenRows) {
-        this.viewportTopR.scrollLeft = x;
-      }
-      this.headerRowScrollerR.scrollLeft = x; // right header row scrolling with frozen grid
-    } else {
-      if (this.freeze.hasFrozenRows) {
-        this.viewportTopL.scrollLeft = x;
-      }
-      this.headerRowScrollerL.scrollLeft = x; // left header row scrolling with regular grid
-    }
-
-    // the bottom-frozen band's scrollable-column viewport follows X like the
-    // frozen-top viewports do
-    if (this.hasBottomFrozenBand()) {
-      (this.hasFrozenColumns() ? this.viewportBottomFrozenR : this.viewportBottomFrozenL).scrollLeft = x;
-    }
-  }
-
-  /** Mirrors the Y scroll position onto the frozen-band viewports that follow the scroll owner. */
-  syncVerticalFollowers(scrollTop: number) {
-    if (this.hasFrozenColumns()) {
-      if (this.freeze.hasFrozenRows && !this.freeze.frozenBottom) {
-        this.viewportBottomL.scrollTop = scrollTop;
-      } else {
-        this.viewportTopL.scrollTop = scrollTop;
-      }
-    }
-
-    // the right-frozen band's scrollable-body viewport follows Y the same way
-    if (this.bands.frozenRightCols > 0 && this.viewportTopRF) {
-      if (this.freeze.hasFrozenRows && !this.freeze.frozenBottom) {
-        this.viewportBottomRF.scrollTop = scrollTop;
-      } else {
-        this.viewportTopRF.scrollTop = scrollTop;
-      }
-    }
-  }
-
-  selectScrollContainers(): { x: HTMLDivElement; y: HTMLDivElement; header: HTMLDivElement; headerRow: HTMLDivElement; footerRow: HTMLDivElement; } {
-    let x: HTMLDivElement;
-    let y: HTMLDivElement;
-    let header: HTMLDivElement;
-    let headerRow: HTMLDivElement;
-    let footerRow: HTMLDivElement;
-
-    if (this.hasFrozenColumns()) {
-      header = this.headerScrollerR;
-      headerRow = this.headerRowScrollerR;
-      footerRow = this.footerRowScrollerR;
-
-      if (this.freeze.hasFrozenRows) {
-        if (this.freeze.frozenBottom) {
-          x = this.viewportBottomR;
-          y = this.viewportTopR;
-        } else {
-          x = y = this.viewportBottomR;
-        }
-      } else {
-        x = y = this.viewportTopR;
-      }
-    } else {
-      header = this.headerScrollerL;
-      headerRow = this.headerRowScrollerL;
-      footerRow = this.footerRowScrollerL;
-
-      if (this.freeze.hasFrozenRows) {
-        if (this.freeze.frozenBottom) {
-          x = this.viewportBottomL;
-          y = this.viewportTopL;
-        } else {
-          x = y = this.viewportBottomL;
-        }
-      } else {
-        x = y = this.viewportTopL;
-      }
-    }
-
-    this.scrollContainers = { x, y, header, headerRow, footerRow };
-    return this.scrollContainers;
-  }
 }
 
 export class SlickGrid<TData = any, C extends Column<TData> = Column<TData>, O extends BaseGridOption<C> = BaseGridOption<C>> {
@@ -1896,7 +387,7 @@ export class SlickGrid<TData = any, C extends Column<TData> = Column<TData>, O e
   protected dragReplaceEl = new DragExtendHandle(this.uid);
   protected _focusSink!: HTMLDivElement;
   protected _focusSink2!: HTMLDivElement;
-  protected _viewportMgr!: ViewportMgr;
+  protected _viewportMgr!: ViewportMgr_;
   /**
    * True once finishInitialization has run its array-wide bindPaneEvents pass.
    * Band materializers bind their new elements only AFTER this point; during the
